requirements.txt
openpyxl==3.1.2
fpdf2==2.8.1
arabic-reshaper==3.0.0
python-bidi==0.6.0
pdfplumber==0.11.8
python-telegram-bot==20.7
python-dotenv==1.0.1
